Add bug branches to contribution guideline
This commit adds a new prefix type to the branching section of the contribution guideline. The feature prefix alone is not sufficient, as it suggests the development of new functionality. In reaction to a "Bug Report" issue, developers can now create branches with the prefix "bug/", to address a problem.
Closes #18