Define Enumeration Characteristic as Named Concept
Since the Characteristic element is not defined as a Named Concept,
Characteristics may be defined as anonymous nodes. Since the bamm:name
attribute has been removed, anonymous nodes do not have a name any more.

This presents a problem regarding Enumerations and code generation. When
loading an Aspect Model, the corresponding implementation needs to
create a synthetic name for all elements defined as anonymous nodes. In
case of an Enumeration this synthetic name would be used when generating
code from the Aspect model, resulting in some random name for the
generated Enum class. For developers using the generated code, classes
with random names which do not depict the purpose of the class is not
particularly satisfying.

Hence an Enumeration is now defined as a Named Concept to ensure that
Enumeration elements have well defined names.

fixes #50